+++ Makefile
@@ -55,13 +55,20 @@ help:
 	@echo "3. In the third terminal window, run:"
 	@echo "    make start-watcher"
 	@echo ""
-	@echo "4. Wait until they all boot. And run in the fourth terminal window:"
+	@echo ""
+	@echo "4. In the third terminal window, run:"
+	@echo "    make start-watcher_info"
+	@echo ""
+	@echo "5. Wait until they all boot. And run in the fourth terminal window:"
 	@echo "    make get-alarms"
 	@echo ""
 	@echo "If you want to attach yourself to running services, use:"
 	@echo "    make remote-child_chain"
 	@echo "or"
 	@echo "    make remote-watcher"
+	@echo ""
+	@echo "or"
+	@echo "    make remote-watcher_info"
 	@echo ""
 	@echo "MISCELLANEOUS"
 	@echo "-------------"
