+++ priv/apps/itest/lib/account.ex
@@ -111,7 +111,7 @@ defmodule Itest.Account do
   # taken from the plasma-contracts deployment snapshot
   # this parsing occurs in several places around the codebase
   defp parse_contracts() do
-    local_umbrella_path = Path.join([File.cwd!(), "../", "localchain_contract_addresses.env"])
+    local_umbrella_path = Path.join([File.cwd!(), "../../../", "localchain_contract_addresses.env"])
 
     contract_addreses_path =
       case File.exists?(local_umbrella_path) do
