+++ priv/apps/itest/mix.exs
@@ -5,7 +5,7 @@ defmodule Itest.MixProject do
     [
       app: :itest,
       version: "0.1.0",
-      elixir: "~> 1.9",
+      elixir: "~> 1.8",
       start_permanent: Mix.env() == :prod,
       deps: deps(),
       build_path: "../../_build",
@@ -25,6 +25,7 @@ defmodule Itest.MixProject do
   # Run "mix help deps" to learn about dependencies.
   defp deps do
     [
+      {:ex_plasma, git: "https://github.com/omisego/ex_plasma.git"},
       {:watcher_info_api, in_umbrella: true},
       {:watcher_security_critical_api, in_umbrella: true},
       {:child_chain_api, in_umbrella: true},
@@ -37,7 +38,7 @@ defmodule Itest.MixProject do
       {:libsecp256k1, git: "https://github.com/omisego/libsecp256k1.git", branch: "elixir-only", override: true},
       {:poison, "~> 3.0"},
       {:tesla, "~> 1.2"},
-      {:white_bread, "~> 4.5.0", only: [:dev, :test]}
+      {:cabbage, "~> 0.3.0"}
     ]
   end
 end
