+++ priv/apps/itest/test/features/in_flight_exits.feature
@@ -1,6 +1,4 @@
 Feature: In Flight Exits
-  # NB: Since the in flight transaction doesn't exist in the system
-  # you don't actually exit any money.
   Scenario: Alice starts an In Flight Exit
     When Alice deposits "1" ETH to the network
     Then Alice should have "1" ETH on the network after finality margin
