--- priv/features/config.exs
@@ -1,10 +1,0 @@
-defmodule WhiteBreadConfig do
-  use WhiteBread.SuiteConfiguration
-
-  suite(
-    name: "All",
-    context: WhiteBreadContext,
-    feature_paths: ["features/"],
-    run_async: false
-  )
-end
